+++ src/gltf/gltf-texture.ts
@@ -1,0 +1,10 @@
+import { Texture } from "pixi.js"
+
+export interface glTFTexture extends Texture {
+  texCoord?: number
+  transform?: {
+    offset?: [number, number]
+    rotation?: number
+    scale?: [number, number]
+  }
+}