+++ src/math/mat3.ts
@@ -1,7 +1,7 @@
 import { mat3 } from "gl-matrix"
 
 export class Mat3 {
-  static multiply(a: Float32Array, b: Float32Array, out: Float32Array) {
+  static multiply(a: Float32Array, b: Float32Array, out = new Float32Array(9)) {
     return <Float32Array>mat3.multiply(out, a, b)
   }
 }