+++ src/texture/khrTextureTransform.ts
@@ -1,6 +1,0 @@
-export interface KHR_texture_transform {
-  offset?: number[]
-  scale?: number[]
-  rotation?: number
-  texCoord?: number
-}